rag_service_pathway/requirements.txt
flask
gunicorn
requests
beautifulsoup4
readability-lxml
python-dotenv
pathway
sentence-transformers>=2.2.2
torch>=1.12.0
faiss-cpu
numpy
tqdm
